# Improve text layer perf