#### Background

TextLayer (100k) example in layer-browser:

| method | time (old) | time (new) |
| -- | -- | -- |
| `transformStringToLetters` | 32585ms | 78ms |

#### Change List
- Remove spread operator construction of arrays
- Avoid merging objects